ci: Improve CI linting  (#791)
* ci: enable more linters


* ci: diable few checks for temporary


* ci: read permission to ci


* ci: try removing cache


* ci: diable ifelse linter for temp


* syle: run gofumpt


* sytle: run gci


* ci: prevent misspell


* ci: run go import


* style: fix style errors


* style: fix trailing lines


* remove unwanted conversion


* remove gci from generated file


* style: remove init check


* fix: type conversion


* fix: remove context


* ci: update configuration file


* ci: skip generated files


* ci: remove disabled lintesr


---------

+++ controllers/redisreplication_controller.go
@@ -26,7 +26,6 @@ type RedisReplicationReconciler struct {
 }
 
 func (r *RedisReplicationReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
-
 	reqLogger := r.Log.WithValues("Request.Namespace", req.Namespace, "Request.Name", req.Name)
 	reqLogger.Info("Reconciling opstree redis replication controller")
 	instance := &redisv1beta2.RedisReplication{}
@@ -73,7 +72,7 @@ func (r *RedisReplicationReconciler) Reconcile(ctx context.Context, req ctrl.Req
 	}
 
 	// Check that the Leader and Follower are ready in redis replication
-	if int32(redisReplicationInfo.Status.ReadyReplicas) != totalReplicas {
+	if redisReplicationInfo.Status.ReadyReplicas != totalReplicas {
 		reqLogger.Info("Redis replication nodes are not ready yet", "Ready.Replicas", strconv.Itoa(int(redisReplicationInfo.Status.ReadyReplicas)), "Expected.Replicas", totalReplicas)
 		return ctrl.Result{RequeueAfter: time.Second * 60}, nil
 	}
@@ -86,12 +85,9 @@ func (r *RedisReplicationReconciler) Reconcile(ctx context.Context, req ctrl.Req
 		if err != nil {
 			return ctrl.Result{RequeueAfter: time.Second * 60}, err
 		}
-
 	}
-
 	reqLogger.Info("Will reconcile redis operator in again 10 seconds")
 	return ctrl.Result{RequeueAfter: time.Second * 10}, nil
-
 }
 
 // SetupWithManager sets up the controller with the Manager.
